requirements.txt
Django==1.10
gunicorn==19.6.0
olefile==0.43
Pillow==4.0.0
whitenoise==3.2.1
dj-database-url==0.4.2
psycopg2==2.6.2
django-rest-framework-social-oauth2==1.0.5
django-bootstrap3==7.0.1
djangorestframework==3.4.3
pytz==2016.6.1
stripe==1.37.0
django-haystack==2.6.0
elasticsearch==5.1.0
drf-haystack==1.6.1
geopy==1.11.0

django-storages==1.5.2
s3transfer==0.1.10
docutils==0.13.1
boto==2.45.0
boto3==1.4.4
botocore==1.5.2
defusedxml==0.4.1
django-braces==1.10.0
django-oauth-toolkit==0.11.0
oauthlib==1.1.2
PyJWT==1.4.2
python-social-auth==0.3.3
python3-openid==3.0.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
